requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
pydantic==1.9.1
PyYAML==6.0
requests==2.28.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
